fix: improve Slider exports and fix slider transition perf
